refactor(#18)+feat: extract parallelBatchFit helper; add to all 9 distributions
Issue #18: Gaussian and Exponential parallelBatchFit were textually
identical (PMD CPD confirmed). Uniform, Discrete, Poisson, and Gamma
had a lighter copy with silent exception swallowing (future.wait()
instead of future.get(), no per-chunk error recording).

Fix: introduce include/core/parallel_batch_fit.h with a single
detail::batchFitParallel<DistT>(datasets, results) function template
that implements the 'heavy' error-recording pattern (pool-access mutex,
atomic has_error, serial fallback). All six existing implementations
are replaced with a one-line delegation.

Three missing distributions (Chi-squared, Student's t, Beta) lacked
parallelBatchFit entirely, violating the 9-distribution standard API.
Add declarations to each header (section 6) and implementations in
the corresponding .cpp files, all delegating to the shared helper.

All 23 correctness tests pass after the change.
