+++ src/beta.cpp
@@ -1,4 +1,5 @@
 #include "libstats/distributions/beta.h"
+#include "libstats/core/parallel_batch_fit.h"
 
 #include "libstats/common/cpu_detection_fwd.h"
 #include "libstats/core/dispatch_utils.h"
@@ -482,6 +483,12 @@ void BetaDistribution::fit(const std::vector<double>& values) {
     }
 
     setParameters(alpha_cur, beta_cur);
+}
+
+void BetaDistribution::parallelBatchFit(
+    const std::vector<std::vector<double>>& datasets,
+    std::vector<BetaDistribution>& results) {
+    detail::batchFitParallel(datasets, results);
 }
 
 void BetaDistribution::reset() noexcept {
