+++ src/discrete.cpp
@@ -1,4 +1,5 @@
 #include "libstats/distributions/discrete.h"
+#include "libstats/core/parallel_batch_fit.h"
 
 // Core functionality - lightweight headers
 #include "libstats/core/dispatch_thresholds.h"
@@ -637,52 +638,10 @@ void DiscreteDistribution::fit(const std::vector<double>& values) {
     cacheValidAtomic_.store(false, std::memory_order_release);
 }
 
-void DiscreteDistribution::parallelBatchFit(const std::vector<std::vector<double>>& datasets,
-                                            std::vector<DiscreteDistribution>& results) {
-    if (datasets.empty()) {
-        // Handle empty datasets gracefully
-        results.clear();
-        return;
-    }
-
-    // Ensure results vector has correct size
-    if (results.size() != datasets.size()) {
-        results.resize(datasets.size());
-    }
-
-    const std::size_t num_datasets = datasets.size();
-
-    // Use distribution-specific parallel thresholds for optimal work distribution
-    if (num_datasets >= detail::dispatch_table::BATCH_FIT_MIN) {
-        // Direct parallel execution without internal thresholds - bypass ParallelUtils limitation
-        ThreadPool& pool = ParallelUtils::getGlobalThreadPool();
-        const std::size_t optimal_grain_size = std::max(std::size_t{1}, num_datasets / 8);
-        std::vector<std::future<void>> futures;
-        futures.reserve((num_datasets + optimal_grain_size - 1) / optimal_grain_size);
-
-        for (std::size_t i = 0; i < num_datasets; i += optimal_grain_size) {
-            const std::size_t chunk_end = std::min(i + optimal_grain_size, num_datasets);
-
-            auto future = pool.submit([&datasets, &results, i, chunk_end]() {
-                for (std::size_t j = i; j < chunk_end; ++j) {
-                    results[j].fit(datasets[j]);
-                }
-            });
-
-            futures.push_back(std::move(future));
-        }
-
-        // Wait for all chunks to complete
-        for (auto& future : futures) {
-            future.wait();
-        }
-
-    } else {
-        // Serial processing for small numbers of datasets
-        for (std::size_t i = 0; i < num_datasets; ++i) {
-            results[i].fit(datasets[i]);
-        }
-    }
+void DiscreteDistribution::parallelBatchFit(
+    const std::vector<std::vector<double>>& datasets,
+    std::vector<DiscreteDistribution>& results) {
+    detail::batchFitParallel(datasets, results);
 }
 
 void DiscreteDistribution::reset() noexcept {
