+++ src/student_t.cpp
@@ -1,4 +1,5 @@
 #include "libstats/distributions/student_t.h"
+#include "libstats/core/parallel_batch_fit.h"
 
 #include "libstats/common/cpu_detection_fwd.h"
 #include "libstats/core/dispatch_utils.h"
@@ -387,6 +388,12 @@ void StudentTDistribution::fit(const std::vector<double>& values) {
     }
 
     setNu(nu);
+}
+
+void StudentTDistribution::parallelBatchFit(
+    const std::vector<std::vector<double>>& datasets,
+    std::vector<StudentTDistribution>& results) {
+    detail::batchFitParallel(datasets, results);
 }
 
 void StudentTDistribution::reset() noexcept {
